+++ internal/env/env.go
@@ -14,6 +14,7 @@ const (
 	DefaultHTTPCode  envVariable = "DEFAULT_HTTP_CODE"  // default HTTP response code
 	ShowDetails      envVariable = "SHOW_DETAILS"       // show request details in response
 	ProxyHTTPHeaders envVariable = "PROXY_HTTP_HEADERS" // proxy HTTP request headers list (request -> response)
+	DisableL10n      envVariable = "DISABLE_L10N"       // disable pages localization
 )
 
 // String returns environment variable name in the string representation.